requirements.txt
flask
python-dotenv
yagmail
python-dotenv
